requirements.txt
tensorflow-gpu
rinokeras==1.1.1
numpy
ipython
tqdm
table-logger
pyyaml
biopython
sacred
pandas
toposort
